Fix `colorcolumn` when `linebreak` is on
The color column used to change its position when `linebreak` is on and
the window is resized. The expected behaviour for a color column would
be that it is always on the same character no matter how often the line
wraps.

An own variable was introduced that tracks the amount of characters that
were inserted due to the `linebreak` and/or the `breakindent` option.
This is used to evaluate the position of the color column.

There is still an issue that the end of the screen line gets highlighted
for some reason.