Change expected 'conceal' test results
Two tests that use the `conceal`, `cursorcolumn`, `linebreak`,
`showbreak` and `colorcolumn` feature failed, due to changes in how the
color column is placed.

Previously, the color column used to change its position based on the
window width and did not respect the `showbreak` characters. As the
highlighted character is now always consistent, regardless of the
window's width, the test now reflects this changed behaviour.

Signed-off-by: Matthias Rader <matthias.rader@outlook.com>